+++ src/main/java/org/example/odiya/route/service/TmapRouteClient.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.example.odiya.common.exception.InternalServerException;
 import org.example.odiya.meeting.domain.Coordinates;
 import org.example.odiya.route.config.RouteClientProperties;
-import org.example.odiya.route.domain.ClientType;
+import org.example.odiya.apicall.domain.ClientType;
 import org.example.odiya.route.domain.RouteInfo;
 import org.example.odiya.route.dto.response.TmapDirectionResponse;
 import org.springframework.http.HttpEntity;
